DES-148: Private List modal
If a non-logged-in user tries to access a private list, they would see
this modal over either an empty list (as shown in "private-list.html")
or the home page. In the modal, user can log in. If user has no
account, she can click "sign-up," which closes the current modal and
opens the Sign Up modal. Clicking cancel from the Sign Up modal brings
the user to the home page.